Credits and Acknowledgements
In root `README.md` add the "Credits" and "Acknowledgements" sections.
The "Credits" section is for third party assets and their license.
The "Acknowledgements" section is for people who helped with the assets
found in this repository. (See #21)
Add back-links to PB Forum profiles, wherever possible.

- In **CREDITS**, mention:
    - Silk Icon Set: by Mark James and Timo «Freak» Harter (CC-BY-2.5).
- In **ACKNOWLEDGEMENTS**, mention:
    - Gary «Kale» Willoughby: for the default theme.
    - Timo «Freak» Harter: for the IDE, the Debugger, an more.
    - Fabien Laboureur: for designing the SpiderBasic logo.

Add or update profile links to people mentioned in other README files,
adopting a consistent mentioning style.