Fix selection of binary numbers
When binary numbers were selected by double-clicking, the prefix (%) of
the binary numbers was not also selected.

When a binary number including the prefix (%) was selected, further
occurrences of this binary number were not highlighted in the code.

Side effect:
If the character '%' is used as an operator in the current case and the
variables and numbers are directly adjacent to the operator (without
spaces), the variables/numbers directly adjacent and the operator are
selected.

Bug in IDE - double click to select problem
https://www.purebasic.fr/english/viewtopic.php?f=24&t=74723

double click select binary value including %
https://www.purebasic.fr/english/viewtopic.php?f=3&t=74741