PB Help FR: Fix Broken HTML Table in "Règles de syntaxe générales"
At the beginning of the French PB documentation page "Règles de syntaxe
générales" there's an HTML table providing a sort of Table of Contents
to the page topics. The last row of the table (which is empty) omitted a
table cell, which caused the table borders to display oddly.

This commit adds the missing cell (`<td></td>`) to fix the problem an
make all table borders look good.