Change #PB_Window_ScreenCentered
On Ubuntu when I have the IDE opened on Monitor #1 this tiny window appear on Monitor #2 and when the IDE is opened on Monitor #2 it appear on Monitor #2 as well but not centered. It appear in the middle of both plus an offset to the right so the entire window is on Monitor #2. By experience with other PureBasic projects I have done this successfully to solve this problem and will don't have an impact on other operating system.