Custom Git Hooks for EditorConfig
Add to the repository scripts to install and remove a custom pre-commit
Git hook that will validate all staged files via EClint:

- `git-hook-install.sh`
- `git-hook-remove.sh`

Add the hook script to export-ignore rules in `.gitattributes`.

The advantage of the hook (compared to `vaildate.sh`) is that it will
only validate the staged changed files, not the whole project folder.

Update `CONTRIBUTING.md` with instructions on how to install and use
the new Git hook.

Set to executable some shell scripts that weren't.