Some PB-IDE fixed for macOS (#226)
* macOS NSToolBar sizeMode

The macOS sizeMode function for NSToolBar no longer exists. This means that the icons (png) with the size 16x16 are scaled by the OS.
To make the icons look better, the 16x16 icons are displayed centred on icons 24x24.

* Update ThemeManagement.pb

Fix CenterImage result

* macOS ListIconGadget Column With

The ListIconGadget must be at least as wide as the first column. Otherwise the first column will not be displayed in width.

* Fix Text Color of Diff Files

* MacOS Update Tools Design (DialogManager)

- Limit minimum height of buttons and string gadgets.
- Panel height increased because gadgets are not completely visible.
- ToolBars changed from TitleBar to Window content.

* Toolbar Update

- Fix #PB_ToolBar_Large